Endre navn på ny comparator til å være Comparator
