+++ Digipost.Api.Client.Tests/Digipost.Api.Client.Tests.csproj
@@ -77,10 +77,9 @@
   </Choose>
   <ItemGroup>
     <Compile Include="ByteUtility.cs" />
-    <Compile Include="CompareObjects\Comparator.cs" />
     <Compile Include="CompareObjects\Difference.cs" />
     <Compile Include="CompareObjects\IDifference.cs" />
-    <Compile Include="CompareObjects\NewComparator.cs" />
+    <Compile Include="CompareObjects\Comparator.cs" />
     <Compile Include="Fakes\FakeAutocompleteResponseHandler.cs" />
     <Compile Include="Fakes\FakeIdentificationResponseHandler.cs" />
     <Compile Include="Fakes\FakeMessageResponseHandler.cs" />
